# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Amana Backend API
  version: 1.0.0
  description: |-
    OpenAPI coverage for the live Amana backend routes under `backend/src/routes`.

    Authentication flow:
    1. Call `POST /auth/challenge` with a Stellar public key.
    2. Sign the returned challenge with the wallet.
    3. Exchange the signature at `POST /auth/verify` for a bearer token.
    4. Send `Authorization: Bearer <token>` to protected endpoints.

    JWT semantics enforced by middleware:
    - `iss` must match `JWT_ISSUER` (default `amana`)
    - `aud` must match `JWT_AUDIENCE` (default `amana-api`)
    - `jti` is required and checked against the revocation denylist
    - `nbf` must be present and not be in the future

    Error code references in operation descriptions map to backend `ErrorCode` values in
    `backend/src/errors/errorCodes.ts`. Some route handlers still emit legacy `{ error }`
    payloads while centralized validation/runtime failures return `{ code, message, details }`.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TradeIdPath:
      in: path
      name: id
      required: true
      description: Trade identifier used by the backend route layer.
      schema:
        type: string
      example: 4294967297
    DisputeIdPath:
      in: path
      name: id
      required: true
      description: Trade identifier for the dispute.
      schema:
        type: string
      example: 4294967297
...
    IdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: false
      description: Optional idempotency key cached for mutation requests for up to 24 hours.
      schema:
        type: string
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example:
        error: Unauthorized
    AppErrorResponse:
      type: object
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTH_ERROR
            - DOMAIN_ERROR
            - INFRA_ERROR
            - NOT_FOUND
            - INTERNAL_ERROR
        message:
          type: string
        details:
          oneOf:
            - type: object
            - type: array
              items:
                type: object
      required: [code, message, details]
      example:
        code: VALIDATION_ERROR
        message: Validation failed
        details:
          - path: body.amountUsdc
            message: Invalid amount format
...
    UnsignedXdrResponse:
      type: object
      properties:
        unsignedXdr:
          type: string
      required: [unsignedXdr]
...
    DisputeResponse:
      type: object
      properties:
        id:
          type: integer
        tradeId:
          type: string
        initiator:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, RESOLVED, CLOSED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        trade:
          type: object
          properties:
            buyerAddress:
              type: string
            sellerAddress:
              type: string
            amountUsdc:
              type: string
      required:
        [id, tradeId, initiator, reason, status, createdAt, updatedAt, trade]
    DisputeListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DisputeResponse"
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
          required: [page, limit, total, totalPages]
      required: [items, pagination]
    DisputeTransitionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [UNDER_REVIEW, RESOLVED, CLOSED]
      required: [status]
      example:
        status: UNDER_REVIEW
...
  responses:
    UnauthorizedError:
      description: Missing, invalid, revoked, or semantically invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unauthorized:
              value:
                error: Unauthorized
            revoked:
              value:
                error: "Unauthorized: token has been revoked"
    ForbiddenError:
      description: Authenticated caller is not allowed to perform the action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: Forbidden
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: Trade not found
    LegacyValidationError:
      description: Request payload failed route-level validation.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - $ref: "#/components/schemas/AppErrorResponse"
    InternalError:
      description: Unexpected backend failure.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - $ref: "#/components/schemas/AppErrorResponse"
paths:
...
  /disputes:
    get:
      tags: [Disputes]
      summary: List disputes for mediators
      description: |-
        Returns paginated list of disputes accessible to the authenticated mediator.
        Error code references: `AUTH_ERROR`, `INTERNAL_ERROR`.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW, RESOLVED, CLOSED]
          description: Filter by dispute status
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        "200":
          description: Dispute list returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Caller is not a mediator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /disputes/{id}/transition:
    post:
      tags: [Disputes]
      summary: Transition dispute status
      description: |-
        Mediator-only endpoint to transition a dispute to a new status.
        Valid transitions: OPEN → UNDER_REVIEW → RESOLVED/CLOSED
        Error code references: `AUTH_ERROR`, `DISPUTE_NOT_FOUND`, `DISPUTE_STATUS_TRANSITION_INVALID`, `DISPUTE_STATUS_CONFLICT`, `INTERNAL_ERROR`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DisputeIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeTransitionRequest"
      responses:
        "200":
          description: Dispute status transitioned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeResponse"
        "400":
          $ref: "#/components/responses/LegacyValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Caller is not a mediator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Dispute not found for the trade.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Dispute was modified by another user (optimistic concurrency conflict).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /trades/{id}/dispute:
    post:
      tags: [Trades]
      summary: Initiate a dispute for a trade
      description: |-
        Supports optional `Idempotency-Key`.
        Route validation requires a reason of at least 10 characters and either `category`
        or `categoryId`. The category must match an active dispute category.
        Error code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `DOMAIN_ERROR`, `NOT_FOUND`, `INTERNAL_ERROR`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TradeIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  minLength: 10
                category:
                  type: string
                  description: Active dispute category name.
                categoryId:
                  type: integer
                  minimum: 1
                  description: Active dispute category id.
              required: [reason]
            example:
              reason: Goods arrived damaged at delivery point.
              category: DAMAGE
      responses:
        "200":
          description: Unsigned dispute transaction returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsignedXdrResponse"
        "400":
          $ref: "#/components/responses/LegacyValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...